requirements.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
crispy-bootstrap5==0.6
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-bootstrap-v5==1.0.7
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.0.0
django-rest-auth==0.9.5
django-tables2==2.4.1
djangorestframework==3.12.4
idna==3.3
isort==5.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
